+++ library/sdxl_train_util.py
@@ -99,7 +99,7 @@ def _load_target_model(name_or_path: str, vae_path: Optional[str], model_version
         state_dict = sdxl_model_util.convert_diffusers_unet_state_dict_to_sdxl(unet.state_dict())
         with init_empty_weights():
             unet = sdxl_original_unet.SdxlUNet2DConditionModel()
-        sdxl_model_util._load_state_dict(unet, state_dict, device=device, dtype=weight_dtype)
+        sdxl_model_util._load_state_dict(unet, state_dict, device=device)
         print("U-Net converted to original U-Net")
 
         logit_scale = None
